split loader: general and a square model test
