+++ src/CMakeLists.txt
@@ -432,10 +432,13 @@ if(BUILD_TESTING)
 
     osh_add_exe(simModel2dLoad)
     set(TEST_EXES ${TEST_EXES} simModel2dLoad)
-    test_func(simModel2dLoad_square 1 ./simModel2dLoad
-      ${CMAKE_SOURCE_DIR}/meshes/square.smd)
     will_fail_test_func(simModel2dLoad_3d 1 ./simModel2dLoad
       ${CMAKE_SOURCE_DIR}/meshes/Example_hex.smd)
+
+    osh_add_exe(simModel2dLoad_square)
+    set(TEST_EXES ${TEST_EXES} simModel2dLoad_square)
+    test_func(simModel2dLoad_square 1 ./simModel2dLoad_square
+      ${CMAKE_SOURCE_DIR}/meshes/square.smd)
   endif()
 
   osh_add_exe(load_2d)
